Restructure utils
* Move featurevector to common.patterns since it is closely related to them
* Move visualize_frame to UDF folder since it depends on LiberTEM core
* Complete list of imports not required since they are imported explicitly in
  examples now
* Use UDF instead of job in test
